// src/components/locations/locations.types.ts
export interface Settings {
  background_color: string;
  glyph: string;
  name: string;
  title: string;
  subtitle: string;
  store: '1' | '2' | '3';
  name_1: string;
  image_1: string | null;
  address_1: string;
  name_2: string;
  image_2: string | null;
  address_2: string;
  name_3: string;
  image_3: string | null;
  address_3: string;
}